Move `useMemoInput` import to `src/hooks` directory
